requirements.txt
selenium
webdriver-manager
python-telegram-bot
beautifulsoup4
requests
